Do not crash when internet is disconnected
Context: https://github.com/an-anime-team/an-anime-game-launcher/issues/483

Users has to disconnect the internet, launch the game then connect the
internet in order to start the game. It's better to support launching
the game with internet disconnected.

Problem: Currently the launcher requires internet to be connected
otherwise it crashes as it cannot detect the game version.

Change: Try to detect the game version, but if it fails just assume it
is the latest version.